Prevent unauthorized edit and delete on others' listings

+++ app/Http/Controllers/ListingController.php
@@ -91,6 +91,11 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        // Make sure current user is the owner of current listing
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -122,6 +127,11 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        // Make sure current user is the owner of current listing
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully.');
     }
